+++ CHANGELOG.md
@@ -9,7 +9,7 @@ project uses [Semantic Versioning](https://semver.org/).
 
 ### Changed
 
-- Replace custom config management with `pydantic-settings`
+- Replace custom config code with `pydantic-settings`
 
 ### Internal